refactor: create project form
